Implement automatic keying, making -k optional
This commit makes the previously required "-k" argument optional,
and also adds a new option "-a" for an autokey path, which
defaults to '/run/tofurkey.autokey' (the /run/ part can be
overridden at build time via the make argument rundir=/x).

This mode of operation isn't useful for cluster-wide
synchronization, although it could still be useful if you just
wants periodic smooth key rotations for a singular server (and
with the default /run/ path, if you also don't mind a
discontinuity on reboots).

The real reason to implement this is that now it's at least
possible to usefully launch the daemon given absolutely no
arguments or prior key generation.  This is necessary if a
distribution package wants to install and enable a unit file for
it during package installation and have it work without human
intervention.