Change default interval to 24h
24h as a default seem to make more sense for most, I think, and is
probably what I'd use in production, I was just being overly
cautious about not making it easier to gather keys for an attack
at the default setting, but really that's inevitably a concern at
any reasonable time-range.  This still contains that possibility
to within a reasonable incident window.

This is a "breaking" change for any active deployment relying on
the default interval, since it will re-key everything as the
upgrade rolls out.  That being said, it's early days and AFAIK no
3rd party is yet using this software, and they could always
explicitly set -i before deploying if it were a concern.  Will
note it in NEWS and bump the minor release JIC.