Pedantic fixes to clock_nanosleep error handling
Surprisingly, clock_nanosleep() returns an errno value directly,
rather than setting errno itself.  This is unlike both
clock_gettime() and nanosleep(), which return -1 and set the
actual errno, like most such functions.

I'm removing the EINTR handling while I'm here, because this was a
vestigial remnant of when I tried to make this work with actual
signal handlers earlier on.  As things stand now (no handlers),
we'll never get an EINTR here anyways, we'll just terminate inside
of the sleep.