Lock all memory and prevent coredump at startup
sodium_malloc() already does this for us for allocations of key
buffers, but really it's safer to just do it at startup for the
entire process, in case key material appears elsewhere in the
process address space temporarily (which is likely!).

Note the use of ONFAULT, which makes this less expensive by not
forcing the immediate faulting of all unfaulted pages.  They'd
have to fault in before we stored a secret to them anyways.

The mlockall() is only attempted when EUID==0.  The rationale is:
* Real production usage of this needs to run as root anyways to
  write the keys to procfs and it's explicitly documented that the
  root user has no rlimits on locked memory.
* Test usage by developers routinely invokes as a non-root user
  using the testing-only flag "-P" for an alternate procfs path.
* Attempting mlockall() for non-root invocations would add all the
  complications of failures due to rlimits and whether we should
  handle them fatally and/or try to raise soft rlimits and/or warn
  the user of how to fix it, etc

For the Zig port, this change required defining our own minimal
syscall wrapper for mlockall(), because Zig std currently lacks
it.  Our minimal wrapper doesn't handle the flags correctly for
the sparc, ppc, or alpha architectures.  Will fix when upstreaming
to std.