Stop using sodium_malloc
Using sodium_malloc() for crypto key storage is basically just a
minor insurance policy thing. Its primary benefits are coredump
avoidance, mlock, and some gaurd page type stuff for the narrow
slices of explicit key memory it allocates.  There are probably
copies elsewhere in memory anyways.

Given that we've upgraded to coredump avoidance and mlock for the
whole process in the prior commit, which provides at least those
benefits to *all* process memory now, it doesn't seem worth it to
bother with sodium_malloc() just to guard-page the main copies of
the keys in such a simple codebase.

We will still zero out our explicit copies, at least, and in the C
case we're doing this in even more edge cases than before, at the
trivial cost of doing it redundantly in some cases.