Zig port: Clean up getopt
This removes a bunch of excessive commentary I made along the way.
It was just way too much!  Some logic was rationalized and
collapsed down a bit as well.  Some true doc comments were added
as necessary, including a nice cheat-sheet for porting at the top,
and the function was placed above the iterator struct for the
documentation to flow more-naturally.

Also, removed the getOptErr() and resetOptInd() interfaces, as
they're mostly pointless bloat.  The caller knows the state of
opterr, and the whole "optind = 0" behavior is (a) not actually
standardized by POSIX and (b) easily emulated by just making a new
getopt iterator object.